move chain info structs
